Remove use of `ignore()` from `trace.js`
Part of migrating to mirror types.

Primitives remain tricky as `String.prototype.replace` had to be avoided due to self-tracing triggered via `Symbol.replace` in Chrome and Firefox.

I decided to remove use of `replace` to help logging perf and left a TODO to figure out how to avoid self-tracing in this case in the future.