Refactor fix for static `this` for simplicity and effeciency
Eliminates the need for `mapGlobalProxyToFunction` while achieving the same effect using a locally scoped reference to the native object.

Also allows the entire override to be done in one spot instead of split between the native object and its static methods.

Lastly removed items from `jsTypes` which are not functions since they are already excluded.