Drop use of `__proto__` setter
Using the third parameter of `Reflect.construct` to specify the prototype of a new object makes this unnecessary.

However, `mirrorResult` was not in charge of creating instances so it was fixed by moving into special handling for each type. Now types that need it mirror results by creating a mirrored copy (as opposed to re-wiring the prototype of the returned non-mirror instance).