Avoid growing list of Firefox WeakMap key bugs
Also affects WeakSet. Apparently the bindings for a number of Firefox's native objects prevent them from being used as WeakMap or WeakSet keys.

Fixing by falling back to Map and Set for now, which will leak some memory when running in Firefox. However, most of the objects we put in here are native APIs and wouldn't be GC'd anyway so the actual leakage is likely low.